requirements.txt
python-docx
pygame
pydub
inflect
noisereduce
soundfile
scipy
numpy
coqui-tts
num2words
cx-Freeze